pyproject.toml
[tool.poetry]
name = "vibezen"
version = "0.1.0"
description = "VIBEcoding Enhancement Zen - AI coding quality assurance system"
authors = ["VIBEZEN Team"]
readme = "README.md"
python = "^3.12"
packages = [{include = "vibezen", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.5"
httpx = "^0.26"
jinja2 = "^3.1"
asyncio = "^3.4"
pyyaml = "^6.0"
rich = "^13.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
pytest-cov = "^4.1"
black = "^24.0"
ruff = "^0.1"
mypy = "^1.8"

[tool.poetry.scripts]
vibezen = "vibezen.cli:main"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q --strict-markers"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"